Add the laminas-codingstandard as a development dependency
This change adds the laminas-codingstandard as a development dependency
as that's what the rest of the laminas projects use. I appreciate that
it's a bit of a departure from what the project uses at the moment. But,
for consistency's sake, it seems like a good choice to make, making the
code easier to read when you're working with laminas projects.